bugfixs for wtrflx
wtrflx has replaced salflx in version 2.3 archive files, and getdat.f
converts salflx from older archive files to wtrflx.  There was a sign
error in this conversion that has been fixed.

For compatibility with previous versions, plots of E-P from hycomproc
are positive into the atmosphere (opposite to HYCOM's usual convention).
This was already correct for archives containing wtrflx and, with the
new getdat, it is now correct for older archives containing salflx only.
Note that prior to version 2.3 salflx contained the flux associated
with relaxation to SSS and this would have been part of the E-P plot.
For version 2.3, E-P is from wtrflx and relaxation to SSS is still in
salflx and so is not included in the E-P plot.

archv2data2d and archv2ncdf2d output fields based on wtrflx and they
are all positive into the ocean.  A sign error has been fixed and
buoflx.f has been updated to use wtrflx instead of salflx.