Compression added as an option for NetCDF output
To turn on compression of NetCDF output in archive programs set the
enviroment variable CDF_DEFLATE to a value between 1 and 9.  The value
of deflate is written to the .log file (with the default value of 0
indicating no compression) and the deflation status of any resulting
NetCDF file can be viewed using the -s option to ncdump:

ncdump -hs GLBy0.04_216_2021012112_t0000_ts3z.nc
netcdf GLBy0.04_216_2021012112_t0000_ts3z {

	short water_temp(time, depth, lat, lon) ;
		water_temp:long_name = "Water Temperature" ;
		water_temp:standard_name = "sea_water_temperature" ;
		water_temp:units = "degC" ;
		water_temp:_FillValue = -30000s ;
		water_temp:missing_value = -30000s ;
		water_temp:scale_factor = 0.001f ;
		water_temp:add_offset = 20.f ;
		water_temp:NAVO_code = 15 ;
		water_temp:_Storage = "chunked" ;
		water_temp:_ChunkSizes = 1, 3, 711, 750 ;
		water_temp:_DeflateLevel = 6 ;
		water_temp:_Shuffle = "true" ;
		water_temp:_Endianness = "little" ;

// global attributes:
		:_Format = "netCDF-4 classic model" ;
}

A deflation level of 6 reduced the size of archv2ncdf3z output files by a
factor of 4 but significantly increased the program's run time.