Control NetCDF chunking of the vertical dimension
To select how the vertical dimension is chunked for NetCDF output
in archive programs set the enviroment variable CDF_KCHUNK to a
value between 1 and 999 (say).  If not defined or 0 the default
chunking is used, and a value greater than the vertical dimension
size is set to that size (e.g. 999 selects the full size when
this is less than 1000).  The value of kchunk is written to the
.log file and the chunking of any resulting NetCDF file can be
viewed using the -s option to ncdump.

A chunksize of 1 may be optimal if you expect to access the variable
one level at a time, and a single vertical chunk may be optimal if
you expect to access the full vertical extent at once.