Check for too many MPI tasks
HYCOM-tools/MPI provides MPI versions of some archive and meanstd tools.
These are primarily designed for cases that need more memory than is
available on one node.  The maximum number of MPI tasks is set by the
parameter jqr in mod_dimensions.F, and this is 90 unless manually edited.
Ninty MPI tasks is typically sufficient because we are more concerned
with available memory than the number of cores used.

There is now an error xcstop if the number of MPI tasks is > jqr.  In
such cases it is likely best to use fewer cores (MPI tasks) per node,
but an alternative is to increase jqr in mod_dimensions.F and remake.