Better interpolation for vertical visc & diff
The vertical viscosity and diffusion fields that can optionally be
in an archive file are at the layer interfaces, rather than the
usual layer averages.

Previously, archv2data3z and archv2ncdf3z treated these as standard
tracers but now if they have *exactly* the archive names they are
treated as interface fields.  For example:

   3    'ntracr' = number of tracers (to output, optional with default 0)
  viscty m^2/s  viscosity   viscty
  t-diff m^2/s  t_diffusion t-diff
  s-diff m^2/s  s_diffusion s-diff
4500    'idm   ' = longitudinal array size
4263    'jdm   ' = latitudinal  array size

This can be checked by outputting the surface (z=0.0) which will
now always be zero.

Note that the MPI/archive version does not handle tracers.