Scripts to download and merge ETOPO_2022_v1_15s
The 15-second version of ETOPO_2022 is only publically available
as about 300 small area tiles and HYCOM-tools requires a single
global field in a netCDF file for bathy_15sec and landsea_15sec.

The script etopo2022_all.csh gets the tiles and merges them into
a netCDF file suitable for HYCOM-tools.  It calls three other
scripts to wget, merge and rename the data set and these can be
invoked separately.